Free Magento 2 Extensions
If yours is a Magento 2 store that you want to enhance, install multiple Magento2 extensions and these will extend the built-in functionality. The Magento2 extensions by Dotsquares Stores are designed to boost the website’s working and to provide the users with a better experience. Install these extensions and save time on extensive coding while giving users a great experience at the same time.